application/single_app/requirements.txt
# requirements.txt
azure-monitor-query==1.4.1
opencensus-ext-azure==1.1.15
Flask==2.2.5
Flask-WTF==1.2.1
gunicorn
Werkzeug==3.0.6
requests==2.32.4
openai==1.67
docx2txt==0.8
Markdown==3.3.4
bleach==6.1.0
azure-cosmos==4.3.0
msal==1.31.0
Flask-Session==0.8.0
azure-ai-documentintelligence==1.0.0b4
numpy==2.1.1
scikit-learn==1.5.2
SciPy==1.15.1
joblib==1.4.2
threadpoolctl==3.5.0
azure-search-documents==11.5.1
python-dotenv==0.21.0
azure-ai-formrecognizer==3.3.3
pyjwt==2.9.0
pandas==2.2.3
markdown2==2.5.3
azure-mgmt-cognitiveservices==13.6.0
azure-identity==1.23.0
azure-ai-projects>=1.0.0b7
azure-ai-contentsafety==1.0.0
azure-storage-blob==12.24.1
azure-storage-queue==12.12.0
pypdf==5.3.1
python-docx==1.1.2
flask-executor==1.0.0
PyMuPDF==1.25.3
langchain-text-splitters==0.3.7
beautifulsoup4==4.13.3
openpyxl==3.1.5
xlrd==2.0.1
pillow==11.1.0
ffmpeg-binaries-compat==1.0.1
ffmpeg-python==0.2.0
semantic-kernel>=1.32.1
redis>=5.0,<6.0
